# backend/cmd/app/docs/swagger.yaml
basePath: /
definitions:
...
  normaladmin_backend_pkg_utils_response.ResponseData:
    properties:
      data: {}
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a backend server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend API
  version: "1.0"
paths:
...
  /gam/upload/batch:
    post:
      consumes:
      - multipart/form-data
      description: 同时上传多个文件
      parameters:
      - description: 要上传的文件数组
        in: formData
        name: files[]
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    name:
                      type: string
                    path:
                      type: string
                    url:
                      type: string
                  type: object
              type: object
        "400":
          description: 获取上传文件失败或上传过程出错
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      summary: 批量上传文件
      tags:
      - 文件管理
  /gam/upload/config:
    get:
      consumes:
      - application/json
      description: 获取文件上传相关的配置信息，包括文件大小限制、支持的文件类型等
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: 获取配置失败
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      summary: 获取上传配置
      tags:
      - 文件管理
  /gam/upload/delete:
    post:
      consumes:
      - application/json
      description: 根据文件路径删除已上传的文件
      parameters:
      - description: 文件路径信息
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "400":
          description: 参数错误或删除失败
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      summary: 删除文件
      tags:
      - 文件管理
  /gam/upload/file:
    post:
      consumes:
      - multipart/form-data
      description: 上传单个文件，支持多种文件类型
      parameters:
      - description: 要上传的文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    name:
                      type: string
                    path:
                      type: string
                    url:
                      type: string
                  type: object
              type: object
        "400":
          description: 获取上传文件失败或上传过程出错
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      summary: 通用文件上传
      tags:
      - 文件管理
  /gam/upload/image:
    post:
      consumes:
      - multipart/form-data
      description: 上传单个图片文件，仅支持图片格式（jpg、jpeg、png、gif等）
      parameters:
      - description: 要上传的图片文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
            - properties:
                data:
                  properties:
                    name:
                      type: string
                    path:
                      type: string
                    url:
                      type: string
                  type: object
              type: object
        "400":
          description: 获取上传文件失败或非图片格式
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/normaladmin_backend_pkg_utils_response.ResponseData'
      summary: 图片文件上传
      tags:
      - 文件管理
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"